# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ScreenTinker Public API
  version: 1.9.0
  description: |
    Public, token-scoped REST API for ScreenTinker digital signage.

    ## Authentication
    Every endpoint (except the two unauthenticated device-render endpoints) accepts a
    **scoped personal access token** presented as `Authorization: Bearer st_...`. The
    dashboard's session JWT is also accepted, but this contract targets token consumers.

    ## Scope ladder (read < write < full)
    Tokens carry one of three scopes; higher scopes subsume lower ones:
    - `read`  — all `GET` endpoints.
    - `write` — all resource mutations (`POST` / `PUT` / `PATCH` / `DELETE`).
    - `full`  — operational fleet commands (currently only
      `POST /groups/{id}/command`: reboot / shutdown / screen on/off).

    Each operation advertises its minimum scope via the `x-required-scope` extension
    and restates it in the operation description. A token with insufficient scope is
    rejected with `403`.

    All resources are scoped to the token's workspace; cross-workspace access is not
    available to tokens.
servers:
  - url: /api
...
security:
  - ApiToken: []
  - SessionJWT: []
components:
  securitySchemes:
    ApiToken:
      type: http
      scheme: bearer
      bearerFormat: "st_<token>"
      description: |
        Scoped personal access token. Send as `Authorization: Bearer st_...`.
        The token's scope (`read`, `write`, or `full`) gates which operations
        it may call (see the scope ladder in the API description).
    SessionJWT:
      type: http
      scheme: bearer
      description: |
        Dashboard session JWT. Also accepted on every authenticated endpoint,
        but this spec targets scoped-token consumers.
  schemas:
...
    Device:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        workspace_id:
          type: [string, "null"]
        playlist_id:
          type: [string, "null"]
        layout_id:
          type: [string, "null"]
          description: Assigned layout (null = fullscreen). Recently added.
        timezone:
          type: string
        orientation:
          type: string
        notes:
          type: string
        created_at:
          type: integer
...
    Group:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          description: "#RRGGBB"
        playlist_id:
          type: [string, "null"]
        device_count:
          type: integer
        workspace_id:
          type: [string, "null"]
...
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
      description: Max rows to return.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
      description: Row offset for pagination.
paths:
...
  /assignments/device/{deviceId}:
    get:
      tags: [assignments]
      summary: List a device's playlist items
      description: 'Requires scope: read.'
      x-required-scope: read
      parameters:
        - { name: deviceId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Array of items. }
    post:
      tags: [assignments]
      summary: Add a content/widget item to a device's playlist
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: deviceId, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content_id: { type: string }
                widget_id: { type: string }
                zone_id: { type: string }
                duration_sec: { type: integer }
                sort_order: { type: integer }
      responses:
        '201': { description: Created item. }
...
  /devices:
    get:
      tags: [devices]
      summary: List devices in the current workspace
      description: 'Requires scope: read.'
      x-required-scope: read
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Array of devices.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Device' }
  /devices/{id}:
    get:
      tags: [devices]
      summary: Get a device with telemetry and assignments
      description: 'Requires scope: read.'
      x-required-scope: read
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Device detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Device' }
        '404': { description: Not found. }
    put:
      tags: [devices]
      summary: Update a device
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                notes: { type: string }
                timezone: { type: string }
                orientation: { type: string }
                default_content_id: { type: string }
                layout_id: { type: [string, "null"] }
      responses:
        '200':
          description: Updated device.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Device' }
    delete:
      tags: [devices]
      summary: Delete a device
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: '{ success }.' }

  # ------------------------------------------------------------------ layouts
...
  /layouts/device/{deviceId}:
    put:
      tags: [layouts]
      summary: Assign a layout to a device
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: deviceId, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                layout_id: { type: [string, "null"] }
      responses:
        '200': { description: '{ success }.' }

  # ------------------------------------------------------------------- groups
  /groups:
    get:
      tags: [groups]
      summary: List device groups in the current workspace
      description: 'Requires scope: read.'
      x-required-scope: read
      responses:
        '200':
          description: Array of groups.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Group' }
    post:
      tags: [groups]
      summary: Create a device group
      description: 'Requires scope: write.'
      x-required-scope: write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                color: { type: string }
      responses:
        '201':
          description: Created group.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
  /groups/{id}:
    put:
      tags: [groups]
      summary: Update a group
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                color: { type: string }
      responses:
        '200': { description: Updated group. }
    delete:
      tags: [groups]
      summary: Delete a group (converts group schedules to per-device)
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: '{ success, schedules_converted, devices }.' }
  /groups/{id}/devices:
    get:
      tags: [groups]
      summary: List devices in a group
      description: 'Requires scope: read.'
      x-required-scope: read
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Array of devices. }
    post:
      tags: [groups]
      summary: Add a device to a group
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id]
              properties:
                device_id: { type: string }
      responses:
        '201': { description: '{ success, playlist_id }.' }
  /groups/{id}/devices/{deviceId}:
    delete:
      tags: [groups]
      summary: Remove a device from a group
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: deviceId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: '{ success }.' }
  /groups/{id}/assign-content:
    post:
      tags: [groups]
      summary: Add content to every device in a group
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content_id]
              properties:
                content_id: { type: string }
                duration_sec: { type: integer }
      responses:
        '200': { description: '{ success, devices_updated }.' }
  /groups/{id}/assign-playlist:
    post:
      tags: [groups]
      summary: Assign a playlist to every device in a group
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [playlist_id]
              properties:
                playlist_id: { type: string }
      responses:
        '200': { description: '{ success, devices_updated }.' }
  /groups/{id}/command:
    post:
      tags: [groups]
      summary: Send an operational command to every device in a group
      description: |
        Sends reboot / shutdown / screen on/off / launch / update to all devices
        in the group. Requires scope: full (the highest scope; tokens with only
        read or write are rejected).
      x-required-scope: full
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [screen_on, screen_off, launch, update, reboot, shutdown, enable_kiosk_lockdown, disable_kiosk_lockdown]
                  description: enable_kiosk_lockdown and disable_kiosk_lockdown require workspace admin access (403 otherwise).
                payload: { type: object }
      responses:
        '200': { description: '{ success, sent, offline, total, results }.' }

  # ---------------------------------------------------------------------- pip
...
  /schedules/device/{deviceId}:
    get:
      tags: [schedules]
      summary: Get schedules for a device (device + group level)
      description: 'Requires scope: read.'
      x-required-scope: read
      parameters:
        - { name: deviceId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Array of schedules. }
...
  /walls/{id}/devices:
    put:
      tags: [walls]
      summary: Set the wall's device grid positions (replaces member set)
      description: 'Requires scope: write.'
      x-required-scope: write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [devices]
              properties:
                devices:
                  type: array
                  items:
                    type: object
                    properties:
                      device_id: { type: string }
                      grid_col: { type: integer }
                      grid_row: { type: integer }
                      rotation: { type: integer }
                      canvas_x: { type: number }
                      canvas_y: { type: number }
                      canvas_width: { type: number }
                      canvas_height: { type: number }
      responses:
        '200': { description: Updated wall with devices. }
...